Add Resin. Move to package manager file layout.
Currently “Incremental” provides “incremental programming helpers” for updating and observing state changes.

As we’ve used this, we have written wrappers and helpers to integrate this into an iOS app. We’ve named this wrapper library “Resin”. This provides a way to manage application state. More detailed info is in the (updated) readme.md file.

Co-authored-by: Sommer Panage <panage@apple.com>
Co-authored-by: Teddy Wyly <twyly@apple.com>
Co-authored-by: Julien Grimault <juelz@apple.com>
Co-authored-by: Robb Böhnke <robb@robb.is>
Co-authored-by: Daniel Eggert <deggert@apple.com>