Update `Resin`
This updates `Resin` where we move away from the `Environment` parameter in `Store` and `Middleware`.

We found that in practice, it is easier to deal with injecting the required dependencies at the `Parcel` level. Each Middleware/Reducer that has dependencies can get them passed in when the parcel is constructed.

Co-authored-by: Sommer Panage <panage@apple.com>
Co-authored-by: Teddy Wyly <twyly@apple.com>
Co-authored-by: Julien Grimault <juelz@apple.com>
Co-authored-by: Robb Böhnke <robb@robb.is>
Co-authored-by: Daniel Eggert <deggert@apple.com>